Module google_replicapoolupdater1_beta1::api

Structs

InstanceUpdate
Update of a single instance.

InstanceUpdateError
Errors that occurred during the instance update.

InstanceUpdateErrorErrors
[Output Only] The array of errors encountered while processing this operation.

InstanceUpdateList
Response returned by ListInstanceUpdates method.

Operation
An operation resource, used to manage asynchronous API requests.

OperationError
[Output Only] If errors occurred during processing of this operation, this field will be populated.

OperationErrorErrors
[Output Only] The array of errors encountered while processing this operation.

OperationList
Contains a list of Operation resources.

OperationWarnings
There is no detailed description.

OperationWarningsData
[Output only] Metadata for this warning in key:value format.

Replicapoolupdater
Central instance to access all Replicapoolupdater related resource activities

RollingUpdate
The following represents a resource describing a single update (rollout) of a group of instances to the given template.

RollingUpdateCancelCall
Cancels an update. The update must be PAUSED before it can be cancelled. This has no effect if the update is already CANCELLED.

RollingUpdateError
[Output Only] Errors that occurred during the rolling update.

RollingUpdateErrorErrors
[Output Only] The array of errors encountered while processing this operation.

RollingUpdateGetCall
Returns information about an update.

RollingUpdateInsertCall
Inserts and starts a new update.

RollingUpdateList
Response returned by List method.

RollingUpdateListCall
Lists recent updates for a given managed instance group, in reverse chronological order and paginated format.

RollingUpdateListInstanceUpdateCall
Lists the current status for each instance within a given update.

RollingUpdateMethods
A builder providing access to all methods supported on rollingUpdate resources. It is not used directly, but through the Replicapoolupdater hub.

RollingUpdatePauseCall
Pauses the update in state from ROLLING_FORWARD or ROLLING_BACK. Has no effect if invoked when the state of the update is PAUSED.

RollingUpdatePolicy
Parameters of the update process.

RollingUpdateResumeCall
Continues an update in PAUSED state. Has no effect if invoked when the state of the update is ROLLED_OUT.

RollingUpdateRollbackCall
Rolls back the update in state from ROLLING_FORWARD or PAUSED. Has no effect if invoked when the state of the update is ROLLED_BACK.

ZoneOperationGetCall
Retrieves the specified zone-specific operation resource.

ZoneOperationListCall
Retrieves the list of Operation resources contained within the specified zone.

ZoneOperationMethods
A builder providing access to all methods supported on zoneOperation resources. It is not used directly, but through the Replicapoolupdater hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.