Struct google_replicapoolupdater1_beta1::Operation

pub struct Operation {
    pub status: Option<String>,
    pub insert_time: Option<String>,
    pub warnings: Option<Vec<OperationWarnings>>,
    pub error: Option<OperationError>,
    pub target_id: Option<String>,
    pub target_link: Option<String>,
    pub start_time: Option<String>,
    pub client_operation_id: Option<String>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub zone: Option<String>,
    pub region: Option<String>,
    pub self_link: Option<String>,
    pub operation_type: Option<String>,
    pub http_error_message: Option<String>,
    pub progress: Option<i32>,
    pub end_time: Option<String>,
    pub http_error_status_code: Option<i32>,
    pub status_message: Option<String>,
    pub user: Option<String>,
}

An operation resource, used to manage asynchronous API requests.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

pause rolling updates (response)

get zone operations (response)

rollback rolling updates (response)

resume rolling updates (response)

insert rolling updates (response)

cancel rolling updates (response)

Fields
status: Option<String>

[Output Only] Status of the operation. Can be one of the following: "PENDING", "RUNNING", or "DONE".

insert_time: Option<String>

[Output Only] The time that this operation was requested. This is in RFC 3339 format.

warnings: Option<Vec<OperationWarnings>>

no description provided

error: Option<OperationError>

[Output Only] If errors occurred during processing of this operation, this field will be populated.

target_id: Option<String>

[Output Only] Unique target id which identifies a particular incarnation of the target.

target_link: Option<String>

[Output Only] URL of the resource the operation is mutating.

start_time: Option<String>

[Output Only] The time that this operation was started by the server. This is in RFC 3339 format.

client_operation_id: Option<String>

no description provided

creation_timestamp: Option<String>

[Output Only] Creation timestamp in RFC3339 text format.

id: Option<String>

[Output Only] Unique identifier for the resource; defined by the server.

kind: Option<String>

[Output Only] Type of the resource. Always replicapoolupdater#operation for Operation resources.

name: Option<String>

[Output Only] Name of the resource.

zone: Option<String>

[Output Only] URL of the zone where the operation resides.

region: Option<String>

[Output Only] URL of the region where the operation resides.

self_link: Option<String>

[Output Only] The fully qualified URL for the resource.

operation_type: Option<String>

no description provided

http_error_message: Option<String>

no description provided

progress: Option<i32>

no description provided

end_time: Option<String>

no description provided

http_error_status_code: Option<i32>

no description provided

status_message: Option<String>

[Output Only] An optional textual description of the current status of the operation.

user: Option<String>

no description provided