Struct google_replicapoolupdater1_beta1::RollingUpdate

pub struct RollingUpdate {
    pub status: Option<String>,
    pub description: Option<String>,
    pub instance_group: Option<String>,
    pub instance_template: Option<String>,
    pub action_type: Option<String>,
    pub user: Option<String>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub self_link: Option<String>,
    pub error: Option<RollingUpdateError>,
    pub policy: Option<RollingUpdatePolicy>,
    pub progress: Option<i32>,
    pub old_instance_template: Option<String>,
    pub status_message: Option<String>,
    pub instance_group_manager: Option<String>,
}

The following represents a resource describing a single update (rollout) of a group of instances to the given template.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

pause rolling updates (none)

rollback rolling updates (none)

get rolling updates (response)

resume rolling updates (none)

list rolling updates (none)

insert rolling updates (request)

list instance updates rolling updates (none)

cancel rolling updates (none)

Fields

status

[Output Only] Status of the update. Possible values are: - "ROLLING_FORWARD": The update is going forward. - "ROLLING_BACK": The update is being rolled back. - "PAUSED": The update is temporarily paused (inactive). - "ROLLED_OUT": The update is finished, all instances have been updated successfully. - "ROLLED_BACK": The update is finished, all instances have been reverted to the previous template. - "CANCELLED": The update is paused and no longer can be resumed, undefined how many instances are running in which template.

description

An optional textual description of the resource; provided by the client when the resource is created.

instance_group

Fully-qualified URL of an instance group being updated. Exactly one of instanceGroupManager and instanceGroup must be set.

instance_template

Fully-qualified URL of an instance template to apply.

action_type

Specifies the action to take for each instance within the instance group. This can be RECREATE which will recreate each instance and is only available for managed instance groups. It can also be REBOOT which performs a soft reboot for each instance and is only available for regular (non-managed) instance groups.

user

[Output Only] User who requested the update, for example: user@example.com.

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

id

[Output Only] Unique identifier for the resource; defined by the server.

kind

[Output Only] Type of the resource.

self_link

[Output Only] The fully qualified URL for the resource.

error

[Output Only] Errors that occurred during the rolling update.

policy

Parameters of the update process.

progress

[Output Only] An optional progress indicator that ranges from 0 to 100. There is no requirement that this be linear or support any granularity of operations. This should not be used to guess at when the update will be complete. This number should be monotonically increasing as the update progresses.

old_instance_template

Fully-qualified URL of the instance template encountered while starting the update.

status_message

[Output Only] An optional textual description of the current status of the update.

instance_group_manager

Fully-qualified URL of an instance group manager being updated. Exactly one of instanceGroupManager and instanceGroup must be set.