Struct google_replicapoolupdater1_beta1::RollingUpdatePolicy

pub struct RollingUpdatePolicy {
    pub auto_pause_after_instances: Option<i32>,
    pub instance_startup_timeout_sec: Option<i32>,
    pub max_num_concurrent_instances: Option<i32>,
    pub min_instance_update_time_sec: Option<i32>,
    pub max_num_failed_instances: Option<i32>,
}

Parameters of the update process.

This type is not used in any activity, and only used as part of another schema.

Fields
auto_pause_after_instances: Option<i32>

Number of instances to update before the updater pauses the rolling update.

instance_startup_timeout_sec: Option<i32>

The maximum amount of time that the updater waits for a HEALTHY state after all of the update steps are complete. If the HEALTHY state is not received before the deadline, the instance update is considered a failure.

max_num_concurrent_instances: Option<i32>

The maximum number of instances that can be updated simultaneously. An instance update is considered complete only after the instance is restarted and initialized.

min_instance_update_time_sec: Option<i32>

The minimum amount of time that the updater spends to update each instance. Update time is the time it takes to complete all update actions (e.g. Stop call on Instance resource in Rolling Reboot), reboot, and initialize. If the instance update finishes early, the updater pauses for the remainder of the time before it starts the next instance update.

max_num_failed_instances: Option<i32>

The maximum number of instance updates that can fail before the group update is considered a failure. An instance update is considered failed if any of its update actions (e.g. Stop call on Instance resource in Rolling Reboot) failed with permanent failure, or if the instance is in an UNHEALTHY state after it finishes all of the update actions.