Struct google_replicapoolupdater1_beta1::RollingUpdateMethods

pub struct RollingUpdateMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on rollingUpdate resources. It is not used directly, but through the Replicapoolupdater hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_replicapoolupdater1_beta1 as replicapoolupdater1_beta1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use replicapoolupdater1_beta1::Replicapoolupdater;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Replicapoolupdater::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `cancel(...)`, `get(...)`, `insert(...)`, `list(...)`, `list_instance_updates(...)`, `pause(...)`, `resume(...)` and `rollback(...)`
// to build up your call.
let rb = hub.rolling_updates();

Methods

impl<'a, C, A> RollingUpdateMethods<'a, C, A>

fn pause(&self, project: &str, zone: &str, rolling_update: &str) -> RollingUpdatePauseCall<'a, C, A>

Create a builder to help you perform the following task:

Pauses the update in state from ROLLING_FORWARD or ROLLING_BACK. Has no effect if invoked when the state of the update is PAUSED.

Arguments

project - The Google Developers Console project name.

zone - The name of the zone in which the update's target resides.

rollingUpdate - The name of the update.

fn rollback(&self, project: &str, zone: &str, rolling_update: &str) -> RollingUpdateRollbackCall<'a, C, A>

Create a builder to help you perform the following task:

Rolls back the update in state from ROLLING_FORWARD or PAUSED. Has no effect if invoked when the state of the update is ROLLED_BACK.

Arguments

project - The Google Developers Console project name.

zone - The name of the zone in which the update's target resides.

rollingUpdate - The name of the update.

fn get(&self, project: &str, zone: &str, rolling_update: &str) -> RollingUpdateGetCall<'a, C, A>

Create a builder to help you perform the following task:

Returns information about an update.

Arguments

project - The Google Developers Console project name.

zone - The name of the zone in which the update's target resides.

rollingUpdate - The name of the update.

fn resume(&self, project: &str, zone: &str, rolling_update: &str) -> RollingUpdateResumeCall<'a, C, A>

Create a builder to help you perform the following task:

Continues an update in PAUSED state. Has no effect if invoked when the state of the update is ROLLED_OUT.

Arguments

project - The Google Developers Console project name.

zone - The name of the zone in which the update's target resides.

rollingUpdate - The name of the update.

fn list(&self, project: &str, zone: &str) -> RollingUpdateListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists recent updates for a given managed instance group, in reverse chronological order and paginated format.

Arguments

project - The Google Developers Console project name.

zone - The name of the zone in which the update's target resides.

fn insert(&self, request: RollingUpdate, project: &str, zone: &str) -> RollingUpdateInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Inserts and starts a new update.

Arguments

request - No description provided.

project - The Google Developers Console project name.

zone - The name of the zone in which the update's target resides.

fn list_instance_updates(&self, project: &str, zone: &str, rolling_update: &str) -> RollingUpdateListInstanceUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the current status for each instance within a given update.

Arguments

project - The Google Developers Console project name.

zone - The name of the zone in which the update's target resides.

rollingUpdate - The name of the update.

fn cancel(&self, project: &str, zone: &str, rolling_update: &str) -> RollingUpdateCancelCall<'a, C, A>

Create a builder to help you perform the following task:

Cancels an update. The update must be PAUSED before it can be cancelled. This has no effect if the update is already CANCELLED.

Arguments

project - The Google Developers Console project name.

zone - The name of the zone in which the update's target resides.

rollingUpdate - The name of the update.